requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.7.0
anyio==3.6.2
appnope==0.1.3
APScheduler==3.10.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arize-phoenix-otel==0.6.1
arrow==1.2.3
asteval==0.9.31
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
awscli==1.36.3
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
billiard==4.1.0
black==24.10.0
bleach==6.0.0
boto3==1.26.161
botocore==1.29.161
cachetools==5.3.1
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
dnspython==2.3.0
docutils==0.16
ecdsa==0.18.0
eventlet==0.33.3
exceptiongroup==1.2.2
executing==1.2.0
fastapi==0.115.0
fastapi-pagination==0.12.5
fastjsonschema==2.16.3
filetype==1.2.0
fonttools==4.39.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.10.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
gptcache==0.1.44
greenlet==2.0.2
groq==0.11.0
grpcio==1.68.0
gunicorn==23.0.0
h11==0.14.0
hiredis==2.2.3
httpcore==0.17.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.1
idna==3.4
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jiter==0.7.1
jmespath==1.0.1
joblib==1.3.2
json_repair==0.29.2
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.7.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
kiwisolver==1.4.4
kombu==5.3.1
langchain==0.3.1
langchain-core==0.3.16
langchain-groq==0.2.0
langchain-openai==0.2.1
langchain-text-splitters==0.3.2
langsmith==0.1.142
leetscraper==2.4.3
llama-cloud==0.1.5
llama-index==0.12.1
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.1
llama-index-embeddings-openai==0.3.0
llama-index-indices-managed-llama-cloud==0.6.2
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.1
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.0
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.0
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.15
loguru==0.6.0
markdownify==0.11.6
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
motor==3.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
newrelic==10.3.0
nltk==3.9.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.26.3
oauthlib==3.2.2
openai==1.58.1
openinference-instrumentation==0.1.18
openinference-instrumentation-langchain==0.1.29
openinference-instrumentation-llama-index==3.0.4
openinference-semantic-conventions==0.1.12
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
orjson==3.10.11
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pinecone-client==2.2.1
pipenv==2024.0.1
platformdirs==3.1.1
pluggy==1.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
propcache==0.2.0
protobuf==4.24.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycurl==7.45.2
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.14.0
PyJWT==2.7.0
pymongo==4.9.1
pymongo-auth-aws==1.1.0
pymongoarrow==0.7.0
pymongocrypt==1.6.1
pyparsing==3.0.9
pypdf==5.1.0
PyPDF2==3.0.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jobspy==1.1.69
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
redis==4.6.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
s3transfer==0.6.0
schedule==1.2.2
scikit-learn==1.4.1.post1
seaborn==0.13.2
selenium==4.26.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==1.4.47
stack-data==0.6.2
starlette==0.38.6
stripe==5.5.0
striprtf==0.0.26
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.3.0
tiktoken==0.8.0
tinycss2==1.2.1
tls-client==1.0.1
tomli==2.1.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.27.0
trio-websocket==0.11.1
typing==3.7.4.3
typing-inspect==0.8.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.0.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.30.6
uvloop==0.17.0
v==1
vine==5.0.0
virtualenv==20.26.3
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.12
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
widgetsnbextension==4.0.5
wrapt==1.17.0
wsproto==1.2.0
yarl==1.8.2
zipp==3.21.0
python-jose==3.3.0